+++ changelog.md
@@ -1,7 +1,7 @@
 <!--
  * @Date: 2023-08-07 15:18:41
  * @LastEditors: hwrn hwrn.aou@sjtu.edu.cn
- * @LastEditTime: 2024-12-24 21:32:07
+ * @LastEditTime: 2024-12-26 20:53:34
  * @FilePath: /genome/changelog.md
  * @Description:
 -->
@@ -17,6 +17,8 @@ changelog for genome
   - feat:
     - `rename_filtered_ls_tsv` to rename bin name after filtering
     - `gff.parse` may make things clearer
+      - seperate `extract`
+      - seperate `translate` (as well as alias `_translate`)
 - 0.2.4:
   - feat:
     - `UniRefClu` method to cluster genes